requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
flask>=2.3.0
gradio>=3.40.0
requests>=2.31.0
matplotlib>=3.7.0
scikit-image>=0.21.0
safetensors>=0.3.0
omegaconf>=2.3.0
einops>=0.6.0 